Rework existing chromatic tests and added constants block to existing dropdown docs
