+++ packages/react-components/components/dropdown/package.json
@@ -1,0 +1,25 @@
+{
+    "name": "@orbit-ui/react-dropdown",
+    "version": "0.1.0",
+    "description": "ShareGate Orbit UI React dropdown components",
+    "repository": "https://github.com/gsoft-inc/sg-orbit.git",
+    "author": "Groupe Sharegate inc.",
+    "license": "Apache-2.0",
+    "main": "dist/index.js",
+    "publishConfig": {
+        "access": "public"
+    },
+    "peerDependencies": {
+        "@orbit-ui/css-normalize": "*",
+        "@orbit-ui/tachyons": "*",
+        "lodash": "*",
+        "react": "*",
+        "react-dom": "*"
+    },
+    "dependencies": {
+        "@babel/runtime-corejs3": "*",
+        "@orbit-ui/icons": "6.0.0",
+        "@orbit-ui/react-components-shared": "2.4.0",
+        "prop-types": "*"
+    }
+}
