+++ storybook/stories/semantic-ui/react/dropdown/inline.stories.mdx
@@ -24,8 +24,24 @@ An inlined dropdown can appear inline with other content.
 ## Usage
 
 <WarningMessage>
-    This component have the same feature set as the <a href="/?path=/docs/components-dropdown-standard--default-story">standard Dropdown</a> component.
+    This component have the same feature set as the <a href="/?path=/docs/components-dropdown-standard--default-story">standard dropdown</a> component.
 </WarningMessage>
+
+### Constants
+
+The following constants are used in the examples below:
+
+export const FRIENDS = [
+    { key: "Jenny Hess", text: "Jenny Hess", value: "Jenny Hess" },
+    { key: "Elliot Fu", text: "Elliot Fu", value: "Elliot Fu" }
+];
+
+```javascript dark
+const FRIENDS = [
+    { key: "Jenny Hess", text: "Jenny Hess", value: "Jenny Hess" },
+    { key: "Elliot Fu", text: "Elliot Fu", value: "Elliot Fu" }
+];
+```
 
 ### Default
 
@@ -36,18 +52,7 @@ A default inlined dropdown.
         <Dropdown
             placeholder="select friend"
             inline
-            options={[
-                {
-                    key: "Jenny Hess",
-                    text: "Jenny Hess",
-                    value: "Jenny Hess"
-                },
-                {
-                    key: "Elliot Fu",
-                    text: "Elliot Fu",
-                    value: "Elliot Fu"
-                }
-            ]}
+            options={FRIENDS}
         />
     </Story>
 </Preview>
@@ -61,18 +66,7 @@ An inlined dropdown can be embedded in a block.
         <div>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque vestibulum et<br /> et lacus at euismod <Dropdown
             placeholder="select friend"
             inline
-            options={[
-                {
-                    key: "Jenny Hess",
-                    text: "Jenny Hess",
-                    value: "Jenny Hess"
-                },
-                {
-                    key: "Elliot Fu",
-                    text: "Elliot Fu",
-                    value: "Elliot Fu"
-                }
-            ]}
+            options={FRIENDS}
         /> elementum viverra maximus.</div>
     </Story>
 </Preview>
@@ -87,18 +81,7 @@ An inlined dropdown can be disabled.
             placeholder="select friend"
             inline
             disabled
-            options={[
-                {
-                    key: "Jenny Hess",
-                    text: "Jenny Hess",
-                    value: "Jenny Hess"
-                },
-                {
-                    key: "Elliot Fu",
-                    text: "Elliot Fu",
-                    value: "Elliot Fu"
-                }
-            ]}
+            options={FRIENDS}
         />
     </Story>
 </Preview>
